move websocket upload server to separate file
